# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrailTeam team management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/carpools:
    get:
      operationId: listEventCarpools
      tags: [carpools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Carpool offers with claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarpoolOfferWithClaims"
    post:
      operationId: createCarpoolOffer
      tags: [carpools]
      summary: Offer to drive (post available seats and bike tray slots)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarpoolOfferBody"
      responses:
        "201":
          description: Carpool offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolOffer"
  /carpools/{offerId}:
    patch:
      operationId: updateCarpoolOffer
      tags: [carpools]
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarpoolOfferBody"
      responses:
        "200":
          description: Updated carpool offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolOffer"
    delete:
      operationId: deleteCarpoolOffer
      tags: [carpools]
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /carpools/{offerId}/claims:
    post:
      operationId: claimCarpool
      tags: [carpools]
      summary: Claim a seat and/or bike tray in a carpool
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimCarpoolBody"
      responses:
        "201":
          description: Carpool claim created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolClaim"
  /carpools/{offerId}/claims/{claimId}:
    delete:
      operationId: cancelCarpoolClaim
      tags: [carpools]
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
        - name: claimId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Claim cancelled
  /events/{id}/carpool-requests:
    get:
      operationId: listEventCarpoolRequests
      tags: [carpools]
      summary: List all ride requests for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of carpool requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarpoolRequestWithUsers"
    post:
      operationId: createCarpoolRequest
      tags: [carpools]
      summary: Request a ride for a rider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarpoolRequestBody"
      responses:
        "201":
          description: Carpool request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolRequestWithUsers"
        "409":
          description: Duplicate request
  /carpool-requests/{id}:
    patch:
      operationId: updateCarpoolRequest
      tags: [carpools]
      summary: Edit an open carpool request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarpoolRequestBody"
      responses:
        "200":
          description: Updated carpool request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolRequestWithUsers"
    delete:
      operationId: deleteCarpoolRequest
      tags: [carpools]
      summary: Delete a carpool request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /carpool-requests/{id}/match:
    post:
      operationId: matchCarpoolRequest
      tags: [carpools]
      summary: Driver accepts a ride request and creates a claim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchCarpoolRequestBody"
      responses:
        "200":
          description: Request matched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolRequestWithUsers"
        "403":
          description: Not the offer owner
        "409":
          description: Request already matched

  # ─── TRAILHEADS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [super_admin, coach, parent, student]
        podId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        gender:
          type: string
          nullable: true
        grade:
          type: integer
          nullable: true
        coachCertLevel:
          type: string
          nullable: true
        notificationsEnabled:
          type: boolean
        emailNotifications:
          type: boolean
        smsNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        defaultCarpoolSeats:
          type: integer
          nullable: true
        defaultCarpoolTrays:
          type: integer
          nullable: true
        notificationPreferences:
          $ref: '#/components/schemas/UserNotificationPreferences'
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - isActive
        - notificationsEnabled
        - emailNotifications
        - smsNotifications
        - pushNotifications
        - createdAt
...
    CarpoolOffer:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        driverUserId:
          type: integer
        availableSeats:
          type: integer
        bikeTrayCount:
          type: integer
        departureLocation:
          type: string
          nullable: true
        departureTime:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - eventId
        - driverUserId
        - availableSeats
        - bikeTrayCount
        - createdAt
    CarpoolOfferWithClaims:
      allOf:
        - $ref: "#/components/schemas/CarpoolOffer"
        - type: object
          properties:
            driver:
              $ref: "#/components/schemas/User"
            claims:
              type: array
              items:
                $ref: "#/components/schemas/CarpoolClaimWithUser"
            seatsRemaining:
              type: integer
            bikeTraysRemaining:
              type: integer
          required: [driver, claims, seatsRemaining, bikeTraysRemaining]
    CarpoolClaim:
      type: object
      properties:
        id:
          type: integer
        carpoolOfferId:
          type: integer
        riderUserId:
          type: integer
        needsSeat:
          type: boolean
        needsBikeTray:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, carpoolOfferId, riderUserId, needsSeat, needsBikeTray, createdAt]
    CarpoolClaimWithUser:
      allOf:
        - $ref: "#/components/schemas/CarpoolClaim"
        - type: object
          properties:
            rider:
              $ref: "#/components/schemas/User"
          required: [rider]
    CreateCarpoolOfferBody:
      type: object
      properties:
        availableSeats:
          type: integer
        bikeTrayCount:
          type: integer
        departureLocation:
          type: string
        departureTime:
          type: string
          format: date-time
        notes:
          type: string
      required: [availableSeats, bikeTrayCount]
    UpdateCarpoolOfferBody:
      type: object
      properties:
        availableSeats:
          type: integer
        bikeTrayCount:
          type: integer
        departureLocation:
          type: string
        departureTime:
          type: string
          format: date-time
        notes:
          type: string
    ClaimCarpoolBody:
      type: object
      properties:
        needsSeat:
          type: boolean
        needsBikeTray:
          type: boolean
        notes:
          type: string
      required: [needsSeat, needsBikeTray]
...
    CarpoolRequestWithUsers:
      allOf:
        - $ref: "#/components/schemas/CarpoolRequest"
        - type: object
          properties:
            rider:
              $ref: "#/components/schemas/User"
              nullable: true
            requestedBy:
              $ref: "#/components/schemas/User"
              nullable: true
            matchedOffer:
              nullable: true
              type: object
              properties:
                id:
                  type: integer
                driver:
                  $ref: "#/components/schemas/User"
                  nullable: true
    CreateCarpoolRequestBody:
      type: object
      properties:
        riderUserId:
          type: integer
          nullable: true
          description: If null, defaults to the requesting user
        needsBikeTray:
          type: boolean
        notes:
          type: string
    UpdateCarpoolRequestBody:
      type: object
      properties:
        needsBikeTray:
          type: boolean
        notes:
          type: string
        status:
          type: string
          enum: [cancelled, matched]
          description: Allowed transitions from open - cancelled or matched (requires matchedOfferId)
        matchedOfferId:
          type: integer
          nullable: true
          description: Required when setting status to matched
    MatchCarpoolRequestBody:
      type: object
      properties:
        offerId:
          type: integer
        autoCreated:
          type: boolean
          description: >
            True when the client auto-created this offer specifically for the match.
            When true the matched rider consumes a seat (reducing displayed availability).
            When false (default) the driver is extending their existing offer capacity and
            the displayed seat count remains unchanged.
      required: [offerId]

    # ─── TRAILHEAD ────────────────────────────────────────────────────────────
...